% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandarin_utils.R
\name{clean_pinyin}
\alias{clean_pinyin}
\title{Clean pinyin}
\usage{
clean_pinyin(pinyin)
}
\arguments{
\item{pinyin}{String of pseduo-pinyin to clean up.}
}
\value{
A string.
}
\description{
This function cleans pinyin
}
