% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandarin_utils.R
\name{get_filtered_audio_info}
\alias{get_filtered_audio_info}
\title{Get filtered audio info from options}
\usage{
get_filtered_audio_info(opts)
}
\arguments{
\item{opts}{A list of options}
}
\value{
A data frame of filtered audio info
}
\description{
This function returns a data frame of filtered audio info
}
