% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandarin_utils.R
\name{get_ipa}
\alias{get_ipa}
\title{Get IPA}
\usage{
get_ipa(pinyin)
}
\arguments{
\item{String}{of pinyin to get pronunciation of.}
}
\value{
A string.
}
\description{
This function gets IPA pronunciation
}
