% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandarin_utils.R
\name{play_media}
\alias{play_media}
\title{Play media}
\usage{
play_media(media_file)
}
\arguments{
\item{media_file}{Filename to play}
}
\value{
\code{invisible(NULL)}
}
\description{
Plays media file in vlc
}
