% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mandarin_utils.R
\name{print_table}
\alias{print_table}
\title{Print table}
\usage{
print_table(df)
}
\arguments{
\item{df}{Data frame to print to terminal}
}
\value{
\code{invisible(NULL)}
}
\description{
Plays table to terminal
}
